refactor: use-bullet hook 분리
